refactor remove models
