+++ examples/solana/contexts/accounts/cache.ts
@@ -1,9 +1,8 @@
 import { AccountInfo, Connection, PublicKey } from '@solana/web3.js';
 import { MintInfo } from '@solana/spl-token';
-import { TokenAccount } from '../../models';
 import { EventEmitter } from '../../utils/eventEmitter';
 import { ParsedAccountBase, AccountParser } from '../../oyster/contexts/accounts/types';
-import { deserializeMint } from '../../oyster';
+import { deserializeMint, TokenAccount } from '../../oyster';
 
 export const genericCache = new Map<string, ParsedAccountBase>();
 const mintCache = new Map<string, MintInfo>();
