+++ examples/solana/utils/utils.ts
@@ -1,10 +1,10 @@
 import { MintInfo } from '@solana/spl-token';
 
-import { TokenAccount } from './../models';
 import { PublicKey } from '@solana/web3.js';
 import BN from 'bn.js';
 import { WAD, ZERO } from '../constants';
 import { TokenInfo } from '@solana/spl-token-registry';
+import { TokenAccount } from '../oyster';
 
 export type KnownTokenMap = Map<string, TokenInfo>;
